0723 - 컨트롤러, 쿼리 파일 분리
